+++ src/permission/worker_permission.cc
@@ -1,6 +1,7 @@
 #include "permission/worker_permission.h"
 
 #include <string>
+#include <unordered_map>
 #include <vector>
 
 namespace node {
