+++ test/parallel/test-cli-permission-deny-fs.js
@@ -72,30 +72,6 @@ const path = require('path');
 }
 
 {
-  const tmpPath = path.resolve('/tmp/');
-  const pathWithComma = path.resolve('/other,path/');
-  const { status, stdout } = spawnSync(
-    process.execPath,
-    [
-      '--experimental-permission',
-      '--allow-fs-write', `${tmpPath};${pathWithComma}`, '--permission-fs-path-delimiter=;', '-e',
-      `console.log(process.permission.has("fs"));
-      console.log(process.permission.has("fs.read"));
-      console.log(process.permission.has("fs.write"));
-      console.log(process.permission.has("fs.write", "/tmp/"));
-      console.log(process.permission.has("fs.write", "/other,path/"));`,
-    ]
-  );
-  const [fs, fsIn, fsOut, fsOutAllowed1, fsOutAllowed2] = stdout.toString().split('\n');
-  assert.strictEqual(fs, 'false');
-  assert.strictEqual(fsIn, 'false');
-  assert.strictEqual(fsOut, 'false');
-  assert.strictEqual(fsOutAllowed1, 'true');
-  assert.strictEqual(fsOutAllowed2, 'true');
-  assert.strictEqual(status, 0);
-}
-
-{
   const { status, stdout } = spawnSync(
     process.execPath,
     [
